HV-1973 - tests refactored
